Update comment flow. Update custom exceptions.

+++ src/Application/YellowMark.AppServices/Comments/Services/CommentService.cs
@@ -1,11 +1,13 @@
 ﻿using AutoMapper;
+using YellowMark.AppServices.Accounts.Services;
+using YellowMark.AppServices.Comments.Exceptions;
 using YellowMark.AppServices.Comments.Repositories;
 using YellowMark.AppServices.Comments.Specifications;
 using YellowMark.AppServices.Specifications;
-using YellowMark.Contracts;
 using YellowMark.Contracts.Comments;
 using YellowMark.Contracts.Pagination;
 using YellowMark.Domain.Comments.Entity;
+using YellowMark.Domain.UserRoles;
 
 namespace YellowMark.AppServices.Comments.Services;
 
@@ -13,23 +15,34 @@ namespace YellowMark.AppServices.Comments.Services;
 public class CommentService : ICommentService
 {
     private readonly ICommentRepository _commentRepository;
+    private readonly IAccountService _accountService;
     private readonly IMapper _mapper;
 
     /// <summary>
     /// Init <see cref="CommentService"/> instance.
     /// </summary>
     /// <param name="commentRepository">Comment repository <see cref="ICommentRepository"/></param>
+    /// <param name="accountService">Account service <see cref="IAccountService"/></param>
     /// <param name="mapper">Ads mapper.</param>
-    public CommentService(ICommentRepository commentRepository, IMapper mapper)
+    public CommentService(
+        ICommentRepository commentRepository,
+        IAccountService accountService,
+        IMapper mapper)
     {
         _commentRepository = commentRepository;
+        _accountService = accountService;
         _mapper = mapper;
     }
 
     /// <inheritdoc/>
     public async Task<Guid> AddCommentAsync(CreateCommentRequest request, CancellationToken cancellationToken)
     {
+        var currentUserInfo = await _accountService.GetAccountInfoAssync(cancellationToken);
+        CommentOperationException.ThrowIfNull(currentUserInfo, "Could not get user info for comment creation.");
+
         var entity = _mapper.Map<CreateCommentRequest, Comment>(request);
+        entity.UserId = currentUserInfo.Id;
+
         await _commentRepository.AddAsync(entity, cancellationToken);
         return entity.Id;
     }
@@ -57,7 +70,16 @@ public class CommentService : ICommentService
     /// <inheritdoc/>
     public async Task<CommentDto> UpdateCommentAsync(Guid id, CreateCommentRequest request, CancellationToken cancellationToken)
     {
+        var currentUserInfo = await _accountService.GetAccountInfoAssync(cancellationToken);
+        CommentOperationException.ThrowIfNull(currentUserInfo, "Could not get user info for comment creation.");
+
         var currentEntity = await _commentRepository.GetByIdAsync(id, cancellationToken);
+        CommentNotFoundException.ThrowIfNull(currentEntity, "Could not get comment for update.");
+
+        if (currentUserInfo.Id != currentEntity.UserId)
+        {
+            throw new CommentPermissionsDeniedException("Current user can not update this comment.");
+        }
 
         var updatedEntity = _mapper.Map<CreateCommentRequest, Comment>(request);
         updatedEntity.Id = id;
@@ -67,10 +89,24 @@ public class CommentService : ICommentService
 
         return _mapper.Map<Comment, CommentDto>(updatedEntity);
     }
-    
+
     /// <inheritdoc/>
     public async Task DeleteCommentByIdAsync(Guid id, CancellationToken cancellationToken)
     {
+        var currentUserInfo = await _accountService.GetAccountInfoAssync(cancellationToken);
+        CommentOperationException.ThrowIfNull(currentUserInfo, "Could not get user info for comment creation.");
+
+        var currentEntity = await _commentRepository.GetByIdAsync(id, cancellationToken);
+        CommentNotFoundException.ThrowIfNull(currentEntity, "There is nothing to delete.");
+
+        if (
+            currentUserInfo.Id != currentEntity.UserId ||
+            currentUserInfo.UserRoles.Contains(UserRoles.Admin) ||
+            currentUserInfo.UserRoles.Contains(UserRoles.SuperUser))
+        { 
+            throw new CommentPermissionsDeniedException("Current user can not delete this comment.");
+        }
+
         await _commentRepository.DeleteAsync(id, cancellationToken);
     }
 }
